+++ src/IRremote.hpp
@@ -313,7 +313,6 @@
 #  if defined(DECODE_DISTANCE_WIDTH)     // universal decoder for pulse distance width protocols - requires up to 750 bytes additional program memory
 #include <ir_DistanceWidthProtocol.hpp>
 #  endif
-
 #endif // #if !defined(USE_IRREMOTE_HPP_AS_PLAIN_INCLUDE)
 
 /**