+++ src/TinyIR.h
@@ -2,7 +2,7 @@
  *  TinyIR.h
  *
  *
- *  Copyright (C) 2021-2022  Armin Joachimsmeyer
+ *  Copyright (C) 2021-2023  Armin Joachimsmeyer
  *  armin.joachimsmeyer@gmail.com
  *
  *  This file is part of IRMP https://github.com/IRMP-org/IRMP.
@@ -34,10 +34,13 @@
  * @{
  */
 
-#define VERSION_IRTINY "1.0.0"
+#define VERSION_IRTINY "1.1.0"
 #define VERSION_IRTINY_MAJOR 1
-#define VERSION_IRTINY_MINOR 0
+#define VERSION_IRTINY_MINOR 1
 #define VERSION_IRTINY_PATCH 0
+
+
+//#define SUPPORT_ONKYO_16_BIT_COMMAND // If activated take the 16 bit (NEC) command as one 16 bit value and not as 8 bit data and 8 bit inverted data
 
 /**
  * Timing for NEC protocol
@@ -65,54 +68,69 @@
 #define NEC_MAXIMUM_REPEAT_DISTANCE (NEC_REPEAT_PERIOD - NEC_MINIMAL_DURATION + 10000) // 70 ms
 
 /**
- * FAST_8_BIT_CS Protocol characteristics:
+ * FAST protocol characteristics:
  * - Bit timing is like JVC
  * - The header is shorter, 4000 vs. 12500
  * - No address and 16 bit data, interpreted as 8 bit command and 8 bit inverted command,
  *     leading to a fixed protocol length of (7 + (16 * 2) + 1) * 526 = 40 * 560 = 21040 microseconds or 21 ms.
- * - Repeats are sent as complete frames but in a 50 ms period.
+ * - Repeats are sent as complete frames but in a 50 ms period / with a 29 ms distance.
  */
-#define FAST_8_BIT_PARITY_ADDRESS_BITS          0 // No address
-#define FAST_8_BIT_PARITY_COMMAND_BITS         16 // Command and inverted command
-#define FAST_8_BIT_PARITY_BITS                 (FAST_8_BIT_PARITY_ADDRESS_BITS + FAST_8_BIT_PARITY_COMMAND_BITS)
+#define FAST_PARITY_ADDRESS_BITS          0 // No address
+#define FAST_PARITY_COMMAND_BITS         16 // Command and inverted command (parity)
+#define FAST_PARITY_BITS                 (FAST_PARITY_ADDRESS_BITS + FAST_PARITY_COMMAND_BITS)
 
-#define FAST_8_BIT_PARITY_UNIT                 526 // 20 periods of 38 kHz (526.315789)
+#define FAST_PARITY_UNIT                 526 // 20 periods of 38 kHz (526.315789)
 
-#define FAST_8_BIT_PARITY_BIT_MARK             FAST_8_BIT_PARITY_UNIT
-#define FAST_8_BIT_PARITY_ONE_SPACE            (3 * FAST_8_BIT_PARITY_UNIT)     // 1578 -> bit period = 2104
-#define FAST_8_BIT_PARITY_ZERO_SPACE           FAST_8_BIT_PARITY_UNIT           //  526 -> bit period = 1052
+#define FAST_PARITY_BIT_MARK             FAST_PARITY_UNIT
+#define FAST_PARITY_ONE_SPACE            (3 * FAST_PARITY_UNIT)     // 1578 -> bit period = 2104
+#define FAST_PARITY_ZERO_SPACE           FAST_PARITY_UNIT           //  526 -> bit period = 1052
 
-#define FAST_8_BIT_PARITY_HEADER_MARK          (4 * FAST_8_BIT_PARITY_UNIT)     // 2104
-#define FAST_8_BIT_PARITY_HEADER_SPACE         (FAST_8_BIT_PARITY_ONE_SPACE)    // 1578
+#define FAST_PARITY_HEADER_MARK          (4 * FAST_PARITY_UNIT)     // 2104
+#define FAST_PARITY_HEADER_SPACE         (FAST_PARITY_ONE_SPACE)    // 1578
 
-#define FAST_8_BIT_PARITY_REPEAT_PERIOD        50000 // Commands are repeated every 50 ms (measured from start to start) for as long as the key on the remote control is held down.
-#define FAST_8_BIT_PARITY_REPEAT_DISTANCE      (FAST_8_BIT_PARITY_REPEAT_PERIOD - (40 * FAST_8_BIT_PARITY_UNIT)) // 29 ms
-#define FAST_8_BIT_PARITY_MAXIMUM_REPEAT_DISTANCE (FAST_8_BIT_PARITY_REPEAT_DISTANCE + 10000) // 47.5 ms
+#define FAST_PARITY_REPEAT_PERIOD        50000 // Commands are repeated every 50 ms (measured from start to start) for as long as the key on the remote control is held down.
+#define FAST_PARITY_REPEAT_DISTANCE      (FAST_PARITY_REPEAT_PERIOD - (40 * FAST_PARITY_UNIT)) // 29 ms
+#define FAST_PARITY_MAXIMUM_REPEAT_DISTANCE (FAST_PARITY_REPEAT_DISTANCE + 10000) // 47.5 ms
 
-#if defined(USE_FAST_8_BIT_AND_PARITY_TIMING)
-#define TINY_ADDRESS_BITS          FAST_8_BIT_PARITY_ADDRESS_BITS
-#define TINY_COMMAND_BITS          FAST_8_BIT_PARITY_COMMAND_BITS
-#define TINY_COMMAND_HAS_8_BIT_PARITY  true
+/*
+ * Definitions to switch between FAST and NEC/ONKYO timing with the same code.
+ */
+#if defined(USE_FAST_PROTOCOL)
+#define TINY_ADDRESS_BITS          FAST_PARITY_ADDRESS_BITS
+#define TINY_COMMAND_BITS          FAST_PARITY_COMMAND_BITS
+#if !defined(TINY_COMMAND_HAS_8_BIT_PARITY)
+#define TINY_COMMAND_HAS_8_BIT_PARITY  true     // 8 bit and 8 bit parity
+//#define TINY_COMMAND_HAS_8_BIT_PARITY  false    //  16 bit command without parity - not tested
+#endif
 
-#define TINY_BITS                  FAST_8_BIT_PARITY_BITS
-#define TINY_UNIT                  FAST_8_BIT_PARITY_UNIT
+#define TINY_BITS                  FAST_PARITY_BITS
+#define TINY_UNIT                  FAST_PARITY_UNIT
 
-#define TINY_HEADER_MARK           FAST_8_BIT_PARITY_HEADER_MARK
-#define TINY_HEADER_SPACE          FAST_8_BIT_PARITY_HEADER_SPACE
+#define TINY_HEADER_MARK           FAST_PARITY_HEADER_MARK
+#define TINY_HEADER_SPACE          FAST_PARITY_HEADER_SPACE
 
-#define TINY_BIT_MARK              FAST_8_BIT_PARITY_BIT_MARK
-#define TINY_ONE_SPACE             FAST_8_BIT_PARITY_ONE_SPACE
-#define TINY_ZERO_SPACE            FAST_8_BIT_PARITY_ZERO_SPACE
+#define TINY_BIT_MARK              FAST_PARITY_BIT_MARK
+#define TINY_ONE_SPACE             FAST_PARITY_ONE_SPACE
+#define TINY_ZERO_SPACE            FAST_PARITY_ZERO_SPACE
 
-#define TINY_MAXIMUM_REPEAT_DISTANCE  FAST_8_BIT_PARITY_MAXIMUM_REPEAT_DISTANCE // for repeat detection
+#define TINY_MAXIMUM_REPEAT_DISTANCE  FAST_PARITY_MAXIMUM_REPEAT_DISTANCE // for repeat detection
 
 #else
 #define ENABLE_NEC_REPEAT_SUPPORT    // Activating this, enables detection of special short frame NEC repeats. Requires 40 bytes program memory.
 
 #define TINY_ADDRESS_BITS          NEC_ADDRESS_BITS // the address bits + parity
+#  if !defined(TINY_ADDRESS_HAS_8_BIT_PARITY)
+#define TINY_ADDRESS_HAS_8_BIT_PARITY  true     // 8 bit and 8 bit parity
+#  endif
+
 #define TINY_COMMAND_BITS          NEC_COMMAND_BITS // the command bits + parity
-#define TINY_ADDRESS_HAS_8_BIT_PARITY  true
-#define TINY_COMMAND_HAS_8_BIT_PARITY  true
+#  if !defined(TINY_COMMAND_HAS_8_BIT_PARITY)
+#    if defined(SUPPORT_ONKYO_16_BIT_COMMAND)
+#define TINY_COMMAND_HAS_8_BIT_PARITY  false    // 16 bit command without parity
+#    else
+#define TINY_COMMAND_HAS_8_BIT_PARITY  true     // 8 bit and 8 bit parity
+#    endif
+#  endif
 
 #define TINY_BITS                  NEC_BITS
 #define TINY_UNIT                  NEC_UNIT
@@ -128,16 +146,17 @@
 #endif
 
 /*
- * This function is called if a complete command was received and must be implemented by the including file (user code)
+ * This function is called if a complete command was received and must be implemented in the file (user code) which includes this library.
+ * The parameter size is dependent of the code variant used in order to save program memory.
  * We have 6 cases: 0, 8 bit or 16 bit address, each with 8 or 16 bit command
  */
 #if (TINY_ADDRESS_BITS > 0)
 #  if TINY_ADDRESS_HAS_8_BIT_PARITY
 // 8 bit address here
 #    if TINY_COMMAND_HAS_8_BIT_PARITY
-extern void handleReceivedTinyIRData(uint8_t aAddress, uint8_t aCommand, uint8_t aFlags);
+extern void handleReceivedTinyIRData(uint8_t aAddress, uint8_t aCommand, uint8_t aFlags);   // Standard NEC callback
 #    else
-extern void handleReceivedTinyIRData(uint8_t aAddress, uint16_t aCommand, uint8_t aFlags);
+extern void handleReceivedTinyIRData(uint8_t aAddress, uint16_t aCommand, uint8_t aFlags); // If SUPPORT_ONKYO_16_BIT_COMMAND is defined
 #    endif
 
 #  else // TINY_ADDRESS_HAS_8_BIT_PARITY
@@ -150,11 +169,11 @@ extern void handleReceivedTinyIRData(uint16_t aAddress, uint16_t aCommand, uint8
 #  endif
 
 #else
-// No address here
+// FAST protocol - No address here
 #  if TINY_COMMAND_HAS_8_BIT_PARITY
-extern void handleReceivedTinyIRData(uint8_t aCommand, uint8_t aFlags);
+extern void handleReceivedTinyIRData(uint8_t aCommand, uint8_t aFlags); // Standard FAST callback
 #  else
-extern void handleReceivedTinyIRData(uint16_t aCommand, uint8_t aFlags);
+extern void handleReceivedTinyIRData(uint16_t aCommand, uint8_t aFlags); // If "TINY_COMMAND_HAS_8_BIT_PARITY  false" is defined. 16 bit without parity.
 #  endif
 #endif
 
@@ -241,10 +260,10 @@ bool initPCIInterruptForTinyReceiver();
 bool enablePCIInterruptForTinyReceiver();
 void disablePCIInterruptForTinyReceiver();
 bool isTinyReceiverIdle();
-#if defined(USE_FAST_8_BIT_AND_PARITY_TIMING)
-void printTinyReceiverResultMinimal(uint16_t aCommand, uint8_t aFlags, Print *aSerial);
+#if defined(USE_FAST_PROTOCOL)
+void printTinyReceiverResultMinimal(Print *aSerial, uint16_t aCommand, uint8_t aFlags);
 #else
-void printTinyReceiverResultMinimal(uint8_t aAddress, uint8_t aCommand, uint8_t aFlags, Print *aSerial);
+void printTinyReceiverResultMinimal(Print *aSerial, uint8_t aAddress, uint8_t aCommand, uint8_t aFlags);
 #endif
 
 void sendFast8BitAndParity(uint8_t aSendPin, uint8_t aCommand, uint_fast8_t aNumberOfRepeats = 0);